+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Handle aborting tests better in the Unity state window.
+- Simplify unnecessarily complex timeout time formatting
 
 ## [4.0.0] - 2021-03-28
 ### Changed