+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.1.0] - 2021-05-13
+
 ### Fixed
 - Fix description of `Return`.
 - Handle source paths not under project gracefully when executing in Unity Editor.
@@ -85,7 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial public release with basic functionality.
 
-[Unreleased]: https://github.com/sbergen/Responsible/compare/v4.0.0...HEAD
+[Unreleased]: https://github.com/sbergen/Responsible/compare/v4.1.0...HEAD
+[4.1.0]: https://github.com/sbergen/Responsible/releases/tag/v4.1.0
 [4.0.0]: https://github.com/sbergen/Responsible/releases/tag/v4.0.0
 [3.0.1]: https://github.com/sbergen/Responsible/releases/tag/v3.0.1
 [3.0.0]: https://github.com/sbergen/Responsible/releases/tag/v3.0.0